+++ extensions/ssh-manager/src/create.tsx
@@ -1,10 +1,20 @@
-import { Action, ActionPanel, Form, showHUD, useNavigation } from "@raycast/api";
+import { Action, ActionPanel, Form, Icon, showHUD, useNavigation } from "@raycast/api";
 import { nanoid } from "nanoid";
 import { getConnections, saveConnections } from "./storage.api";
 import { ISSHConnection } from "./types";
+import { FormValidation, useForm } from "@raycast/utils";
 
 export default function Main() {
   const { pop } = useNavigation();
+
+  const {handleSubmit, itemProps} = useForm<ISSHConnection>({
+    onSubmit(values) {
+      saveConnection(values);
+    },
+    validation: {
+      name: FormValidation.Required
+    }
+  })
 
   async function saveConnection(connection: ISSHConnection) {
     const existingConnections = await getConnections();
@@ -20,23 +30,23 @@ export default function Main() {
     <Form
       actions={
         <ActionPanel>
-          <Action.SubmitForm title="Save" onSubmit={(values: ISSHConnection) => saveConnection(values)} />
+          <Action.SubmitForm icon={Icon.SaveDocument} title="Save" onSubmit={handleSubmit} />
         </ActionPanel>
       }
     >
-      <Form.TextField id="name" title="Connection Name" />
-      <Form.TextField id="address" title="Server Address" placeholder={"A resolvable DNS name or IP"} />
-      <Form.TextField id="user" title="Username (optional)" placeholder={"A username to authenticate with"} />
-      <Form.TextField id="port" title="Port (optional)" placeholder={"An optional custom port (other than 22)"} />
+      <Form.TextField title="Connection Name" {...itemProps.name} />
+      <Form.TextField title="Server Address" placeholder={"A resolvable DNS name or IP"} {...itemProps.address} />
+      <Form.TextField title="Username (optional)" placeholder={"A username to authenticate with"} {...itemProps.user} />
+      <Form.TextField title="Port (optional)" placeholder={"An optional custom port (other than 22)"} {...itemProps.port} />
       <Form.TextField
-        id="sshKey"
         title="SSH Key Location (optional)"
         placeholder={"An optional key path to authenticate with"}
+        {...itemProps.sshKey}
       />
       <Form.TextField
-        id="command"
         title="Command to Execute (optional)"
         placeholder={"An optional command to execute on the remote server after connecting"}
+        {...itemProps.command}
       />
     </Form>
   );